refactor: move instances to file
